Move dispatch handlers into `dispatch` module
