+++ src/wayland/dispatch/buffer.rs
@@ -1,0 +1,30 @@
+/*
+    WlBuffer related dispatch handlers
+*/
+
+use wayland_client::{
+    Connection, Dispatch, QueueHandle,
+    protocol::wl_buffer::{self, WlBuffer},
+};
+
+use crate::wayland::state::LockState;
+
+impl Dispatch<WlBuffer, i32> for LockState {
+    fn event(
+        state: &mut Self,
+        _proxy: &WlBuffer,
+        event: <WlBuffer as wayland_client::Proxy>::Event,
+        data: &i32,
+        _conn: &Connection,
+        _qh: &QueueHandle<Self>,
+    ) {
+        match event {
+            wl_buffer::Event::Release => {
+                if let Some(buffers) = state.interfaces.buffers.as_mut() {
+                    buffers[*data as usize].in_use = false;
+                }
+            }
+            _ => {}
+        }
+    }
+}
