+++ src/wayland/dispatch/mod.rs
@@ -1,1 +1,6 @@
+pub mod buffer;
 pub mod empty;
+pub mod keyboard;
+pub mod lock;
+pub mod output;
+pub mod registry;
