+++ src/wayland/event.rs
@@ -1,11 +1,9 @@
-use std::os::fd::{AsRawFd, BorrowedFd};
-
-use nix::sys::epoll::{Epoll, EpollCreateFlags, EpollEvent, EpollFlags, EpollTimeout};
-use wayland_client::EventQueue;
-use wayland_client::backend::ReadEventsGuard;
-
 use crate::wayland::lock::State;
 use crate::wayland::state::LockState;
+use nix::sys::epoll::{Epoll, EpollCreateFlags, EpollEvent, EpollFlags, EpollTimeout};
+use std::os::fd::{AsRawFd, BorrowedFd};
+use wayland_client::EventQueue;
+use wayland_client::backend::ReadEventsGuard;
 
 const WAYLAND_EVENT_TAG: u64 = 0;
 const RENDERER_EVENT_TAG: u64 = 1;
